// apps/applio-org/src/types/modelsTypes.ts
export type Model = {
    id_: string;
    name: string;
    link: string;
    epochs: string;
    type: string;
    algorithm: string;
    image_url: string;
    id: string;
    author_id: string;
    created_at: string;
    likes: string;
    author_username: string;
    server_id: string;
    server_name: string;
    tags: string;
};